move collectSources to Script; does not expose Script.script
